+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ## UNRELEASED
 
-### Changed
+### Added
 
-- ...
+- Basic features:
+  - Metrics endpoint
+  - Ping endpoint
+  - Allowed routes list (if requested `/`)
+  - Proxy route
+- CLI allows to use options (can read it from environment variables) like listen address, port, and proxy route prefix
 
 [keepachangelog]:https://keepachangelog.com/en/1.0.0/
 [semver]:https://semver.org/spec/v2.0.0.html